requirements.txt
python-telegram-bot[job-queue]==21.6
pdfplumber==0.11.4
requests==2.32.3
gspread==6.1.2
google-auth==2.34.0
python-dotenv==1.0.1
tenacity==9.0.0
